+++ packages/react/src/evm/providers/wallet-ui-states-provider.tsx
@@ -1,8 +1,9 @@
 import { useState, createContext, useContext } from "react";
-import { isMobile } from "../utils/isMobile";
 import { Theme } from "../../design-system";
 import { WelcomeScreen } from "../../wallet/ConnectWallet/screens/types";
 import { useTWLocale } from "./locale-provider";
+import { canFitWideModal } from "../utils/canFitWIdeModal";
+import { WalletInstance } from "@thirdweb-dev/react-core";
 
 type BoolSetter = (value: boolean) => void;
 
@@ -21,7 +22,8 @@ export type ModalConfig = {
     onLogout?: () => void;
   };
   isEmbed?: boolean;
-  onConnect?: () => void;
+  onConnect?: (wallet: WalletInstance) => void;
+  showThirdwebBranding?: boolean;
 };
 
 const WalletModalOpen = /* @__PURE__ */ createContext(false);
@@ -55,18 +57,19 @@ export const WalletUIStatesProvider = (
       onLogin?: (token: string) => void;
       onLogout?: () => void;
     };
-    onConnect?: () => void;
+    onConnect?: (wallet: WalletInstance) => void;
+    showThirdwebBranding?: boolean;
   }>,
 ) => {
   const [isWalletModalOpen, setIsWalletModalOpen] = useState(false);
-  const _isMobile = isMobile();
+  const enforceCompact = !canFitWideModal();
   const locale = useTWLocale();
 
   const [modalConfig, setModalConfig] = useState<ModalConfig>({
     title: props.title || locale.connectWallet.defaultModalTitle,
     theme: props.theme || "dark",
     data: undefined,
-    modalSize: (_isMobile ? "compact" : props.modalSize) || "wide",
+    modalSize: (enforceCompact ? "compact" : props.modalSize) || "wide",
     termsOfServiceUrl: props.termsOfServiceUrl,
     privacyPolicyUrl: props.privacyPolicyUrl,
     welcomeScreen: props.welcomeScreen,
@@ -74,6 +77,7 @@ export const WalletUIStatesProvider = (
     isEmbed: props.isEmbed,
     auth: props.auth,
     onConnect: props.onConnect,
+    showThirdwebBranding: props.showThirdwebBranding,
   });
 
   return (
@@ -233,6 +237,13 @@ export type ModalConfigOptions = {
    * Note that this does not include the sign in, If you want to call a callback after user connects AND signs in with their wallet, use `auth.onLogin` instead
    */
   onConnect?: () => void;
+
+  /**
+   * By default the ConnectWallet Modal shows "powered by thirdweb" branding at the bottom of the modal.
+   *
+   * If you want to hide the branding, set this to `false`
+   */
+  showThirdwebBranding?: boolean;
 };
 
 /**
@@ -405,7 +416,7 @@ export type ModalConfigOptions = {
  */
 export const useSetWalletModalConfig = () => {
   const context = useContext(SetModalConfigCtx);
-  const _isMobile = isMobile();
+  const enforceCompact = !canFitWideModal();
   const locale = useTWLocale();
 
   if (context === undefined) {
@@ -419,7 +430,7 @@ export const useSetWalletModalConfig = () => {
       title: title || locale.connectWallet.defaultModalTitle,
       data: undefined,
       theme: theme || "dark",
-      modalSize: (_isMobile ? "compact" : modalSize) || "wide",
+      modalSize: (enforceCompact ? "compact" : modalSize) || "wide",
       ...rest,
     });
   };