+++ packages/react/src/wallet/ConnectWallet/Modal/ConnectModalInline.tsx
@@ -9,8 +9,7 @@ import {
   modalMaxWidthWide,
 } from "../constants";
 import { ConnectModalContent } from "./ConnectModal";
-import { useScreen } from "./screen";
-import { isMobile } from "../../../evm/utils/isMobile";
+import { useSetupScreen } from "./screen";
 import { useWallets } from "@thirdweb-dev/react-core";
 import { DynamicHeight } from "../../../components/DynamicHeight";
 import {
@@ -20,6 +19,7 @@ import {
 import { StyledDiv } from "../../../design-system/elements";
 import { SyncedWalletUIStates } from "./ConnectEmbed";
 import { WelcomeScreen } from "../screens/types";
+import { canFitWideModal } from "../../../evm/utils/canFitWIdeModal";
 
 /**
  * @internal
@@ -27,67 +27,62 @@ import { WelcomeScreen } from "../screens/types";
 export type ConnectModalInlineProps = {
   className?: string;
   theme?: "dark" | "light" | Theme;
-
-  /**
-   * Set a custom title for the modal
-   * @defaultValue "Connect"
-   */
   modalTitle?: string;
-
-  /**
-   * Replace the thirdweb icon next to modalTitle and set your own iconUrl
-   *
-   * Set to empty string to hide the icon
-   */
   modalTitleIconUrl?: string;
-
   style?: React.CSSProperties;
-
-  /**
-   * Set the size of the modal - `compact` or `wide` on desktop
-   *
-   * Modal size is always `compact` on mobile
-   *
-   * @defaultValue "wide"
-   */
   modalSize?: "compact" | "wide";
-
-  /**
-   * If provided, Modal will show a Terms of Service message at the bottom with below link
-   */
   termsOfServiceUrl?: string;
-
-  /**
-   * If provided, Modal will show a Privacy Policy message at the bottom with below link
-   */
   privacyPolicyUrl?: string;
-
-  /**
-   * Customize the welcome screen
-   *
-   * Either provide a component to replace the default screen entirely
-   *
-   * or an object with title, subtitle and imgSrc to change the content of the default screen
-   */
   welcomeScreen?: WelcomeScreen;
+  showThirdwebBranding?: boolean;
 };
 
 /**
  * @internal
  */
 export const ConnectModalInline = (props: ConnectModalInlineProps) => {
-  const { screen, setScreen, initialScreen } = useScreen();
   const walletConfigs = useWallets();
   const modalSize =
-    isMobile() || walletConfigs.length === 1 ? "compact" : props.modalSize;
+    !canFitWideModal() || walletConfigs.length === 1
+      ? "compact"
+      : props.modalSize;
   const ctxTheme = useCustomTheme();
+
+  const walletUIStatesProps = {
+    theme: props.theme || ctxTheme,
+    modalSize: modalSize,
+    title: props.modalTitle,
+    termsOfServiceUrl: props.termsOfServiceUrl,
+    privacyPolicyUrl: props.privacyPolicyUrl,
+    welcomeScreen: props.welcomeScreen,
+    titleIconUrl: props.modalTitleIconUrl,
+    showThirdwebBranding: props.showThirdwebBranding,
+  };
+
+  return (
+    <WalletUIStatesProvider {...walletUIStatesProps}>
+      <CustomThemeProvider theme={walletUIStatesProps.theme}>
+        <ConnectModalInlineContent
+          className={props.className}
+          modalSize={modalSize}
+        />
+        <SyncedWalletUIStates {...walletUIStatesProps} />
+      </CustomThemeProvider>
+    </WalletUIStatesProvider>
+  );
+};
+
+function ConnectModalInlineContent(props: {
+  className?: string;
+  modalSize?: "compact" | "wide";
+  style?: React.CSSProperties;
+}) {
+  const screenSetup = useSetupScreen();
 
   const content = (
     <>
       <ConnectModalContent
-        initialScreen={initialScreen}
-        screen={screen}
-        setScreen={setScreen}
+        screenSetup={screenSetup}
         onHide={() => {
           // no op
         }}
@@ -115,41 +110,26 @@ export const ConnectModalInline = (props: ConnectModalInlineProps) => {
     </>
   );
 
-  const walletUIStatesProps = {
-    theme: props.theme || ctxTheme,
-    modalSize: modalSize,
-    title: props.modalTitle,
-    termsOfServiceUrl: props.termsOfServiceUrl,
-    privacyPolicyUrl: props.privacyPolicyUrl,
-    welcomeScreen: props.welcomeScreen,
-    titleIconUrl: props.modalTitleIconUrl,
-  };
-
   return (
-    <WalletUIStatesProvider {...walletUIStatesProps}>
-      <CustomThemeProvider theme={walletUIStatesProps.theme}>
-        <ConnectModalInlineContainer
-          className={props.className}
-          style={{
-            height: modalSize === "compact" ? "auto" : wideModalMaxHeight,
-            maxWidth:
-              modalSize === "compact"
-                ? modalMaxWidthCompact
-                : modalMaxWidthWide,
-            ...props.style,
-          }}
-        >
-          {modalSize === "compact" ? (
-            <DynamicHeight> {content} </DynamicHeight>
-          ) : (
-            content
-          )}
-          <SyncedWalletUIStates {...walletUIStatesProps} />
-        </ConnectModalInlineContainer>
-      </CustomThemeProvider>
-    </WalletUIStatesProvider>
+    <ConnectModalInlineContainer
+      className={props.className}
+      style={{
+        height: props.modalSize === "compact" ? "auto" : wideModalMaxHeight,
+        maxWidth:
+          props.modalSize === "compact"
+            ? modalMaxWidthCompact
+            : modalMaxWidthWide,
+        ...props.style,
+      }}
+    >
+      {props.modalSize === "compact" ? (
+        <DynamicHeight> {content} </DynamicHeight>
+      ) : (
+        content
+      )}
+    </ConnectModalInlineContainer>
   );
-};
+}
 
 const ConnectModalInlineContainer = /* @__PURE__ */ StyledDiv(() => {
   const theme = useCustomTheme();