+++ packages/sdk/src/evm/common/feature-detection/isFeatureEnabled.ts
@@ -3,7 +3,6 @@ import {
   FeatureWithEnabled,
 } from "../../constants/contract-features";
 import { AbiInput } from "../../schema/contracts/custom";
-import { detectFeatures } from "./detectFeatures";
 
 /**
  * Checks whether the given ABI supports a given feature
@@ -15,8 +14,8 @@ import { detectFeatures } from "./detectFeatures";
 export function isFeatureEnabled(
   abi: AbiInput,
   featureName: FeatureName,
+  features: Record<string, FeatureWithEnabled>,
 ): boolean {
-  const features = detectFeatures(abi);
   return _featureEnabled(features, featureName);
 }
 
@@ -26,8 +25,12 @@ export function isFeatureEnabled(
  * @param abi - The abi to check
  * @param featureName - The feature name to check
  */
-export function isExtensionEnabled(abi: AbiInput, featureName: FeatureName) {
-  return isFeatureEnabled(abi, featureName);
+export function isExtensionEnabled(
+  abi: AbiInput,
+  featureName: FeatureName,
+  features: Record<string, FeatureWithEnabled>,
+) {
+  return isFeatureEnabled(abi, featureName, features);
 }
 
 /**