+++ packages/sdk/src/evm/constants/urls.ts
@@ -8,9 +8,8 @@ import { getValidChainRPCs } from "@thirdweb-dev/chains";
 import type { Chain } from "@thirdweb-dev/chains";
 import { providers } from "ethers";
 import type { Signer } from "ethers";
-import pkg from "../../../package.json";
-import { isBrowser } from "@thirdweb-dev/storage";
 import { sha256HexSync } from "@thirdweb-dev/crypto";
+import { setAnalyticsHeaders } from "../../core/utils/headers";
 
 /**
  * @internal
@@ -252,6 +251,8 @@ export function getProviderFromRpcUrl(
         headers["x-authorize-wallet"] = "true";
       }
 
+      setAnalyticsHeaders(headers);
+
       const bundleId =
         typeof globalThis !== "undefined" && "APP_BUNDLE_ID" in globalThis
           ? ((globalThis as any).APP_BUNDLE_ID as string)
@@ -259,16 +260,6 @@ export function getProviderFromRpcUrl(
       if (!rpcUrl.includes("bundleId")) {
         rpcUrl = rpcUrl + (bundleId ? `?bundleId=${bundleId}` : "");
       }
-
-      headers["x-sdk-version"] = pkg.version;
-      headers["x-sdk-name"] = pkg.name;
-      headers["x-sdk-platform"] = bundleId
-        ? "react-native"
-        : isBrowser()
-        ? (window as any).bridge !== undefined
-          ? "webGL"
-          : "browser"
-        : "node";
     }
     const match = rpcUrl.match(/^(ws|http)s?:/i);
     // Try the JSON batch provider if available
@@ -284,7 +275,7 @@ export function getProviderFromRpcUrl(
           if (existingProvider) {
             return existingProvider;
           }
-          
+
           // TODO: remove below `skipFetchSetup` logic when ethers.js v6 support arrives
           let _skipFetchSetup = false;
           if (
@@ -292,7 +283,8 @@ export function getProviderFromRpcUrl(
             "TW_SKIP_FETCH_SETUP" in globalThis &&
             typeof (globalThis as any).TW_SKIP_FETCH_SETUP === "boolean"
           ) {
-            _skipFetchSetup = (globalThis as any).TW_SKIP_FETCH_SETUP as boolean;
+            _skipFetchSetup = (globalThis as any)
+              .TW_SKIP_FETCH_SETUP as boolean;
           }
 
           // Otherwise, create a new provider on the specific network
@@ -305,6 +297,7 @@ export function getProviderFromRpcUrl(
                   skipFetchSetup: _skipFetchSetup,
                 },
                 chainId,
+                sdkOptions?.rpcBatchSettings,
               )
             : // Otherwise fall back to the built in json rpc batch provider
               new providers.JsonRpcBatchProvider({