+++ packages/service-utils/src/cf-worker/usage.ts
@@ -23,6 +23,7 @@ const usageEventSchema = z.object({
     "relayer",
     "connectWallet",
     "checkout",
+    "engine",
   ]),
   action: z.string(),
 
@@ -52,6 +53,18 @@ const usageEventSchema = z.object({
   sdkVersion: z.string().optional(),
   sdkPlatform: z.string().optional(),
   productName: z.string().optional(),
+  transactionValue: z.string().optional(),
+  pathname: z.string().optional(),
+  contractAddress: z.string().optional(),
+  errorCode: z.string().optional(),
+  httpStatusCode: z.number().int().nonnegative().optional(),
+  functionName: z.string().optional(),
+  extension: z.string().optional(),
+  retryCount: z.number().int().nonnegative().optional(),
+  policyId: z.string().optional(),
+  msSinceQueue: z.number().nonnegative().optional(),
+  msSinceSend: z.number().nonnegative().optional(),
+  msTotalDuration: z.number().nonnegative().optional(),
 });
 export type UsageEvent = z.infer<typeof usageEventSchema>;